TEST TDD for PERMISSION
Permission: a type which store the permission and provide function to check if a command can be executed by someone. It provides a cache as there is recursive permission to include right.
example: Admin can be included in moderator which can be included in member. Thus admins can do everything a mod do abd everything a member do